SIP-5148: while deleting channel a message asking confirmation should come SIP-5152: Edit channel dialog heading SIP-5154: Delete route dialog

+++ saw-web/src/app/modules/workbench/components/datasource-management/datasource-page.component.ts
@@ -11,7 +11,7 @@ import { ToastService } from '../../../../common/services/toastMessage.service';
 import { CreateSourceDialogComponent } from './createSource-dialog/createSource-dialog.component';
 import { CreateRouteDialogComponent } from './create-route-dialog/create-route-dialog.component';
 import { TestConnectivityComponent } from './test-connectivity/test-connectivity.component';
-
+import { ConfirmActionDialogComponent } from './confirm-action-dialog/confirm-action-dialog.component';
 import * as isUndefined from 'lodash/isUndefined';
 import * as forEach from 'lodash/forEach';
 import * as countBy from 'lodash/countBy';
@@ -159,9 +159,17 @@ export class DatasourceComponent implements OnInit, OnDestroy {
   }
 
   deleteChannel(channelID) {
-    this.datasourceService.deleteChannel(channelID).subscribe(() => {
-      this.notify.success('Channel deleted successfully');
-      this.getSources();
+    const dialogRef = this.dialog.open(ConfirmActionDialogComponent, {
+      width: '350px'
+    });
+
+    dialogRef.afterClosed().subscribe(confirmed => {
+      if (confirmed) {
+        this.datasourceService.deleteChannel(channelID).subscribe(() => {
+          this.notify.success('Channel deleted successfully');
+          this.getSources();
+        });
+      }
     });
   }
 
@@ -246,11 +254,19 @@ export class DatasourceComponent implements OnInit, OnDestroy {
   }
 
   deleteRoute(routeData) {
-    const channelID = routeData.bisChannelSysId;
-    const routeID = routeData.bisRouteSysId;
+    const dialogRef = this.dialog.open(ConfirmActionDialogComponent, {
+      width: '350px'
+    });
 
-    this.datasourceService.deleteRoute(channelID, routeID).subscribe(() => {
-      this.getRoutesForChannel(channelID);
+    dialogRef.afterClosed().subscribe(confirmed => {
+      if (confirmed) {
+        const channelID = routeData.bisChannelSysId;
+        const routeID = routeData.bisRouteSysId;
+
+        this.datasourceService.deleteRoute(channelID, routeID).subscribe(() => {
+          this.getRoutesForChannel(channelID);
+        });
+      }
     });
   }
